refractor(Home): changed induction controller and bindings to home
