+++ lib/bindings/home_binding.dart
@@ -1,0 +1,9 @@
+import 'package:get/instance_manager.dart';
+import 'package:nittfest/controllers/home_controller.dart';
+
+class HomeBindings extends Bindings {
+  @override
+  void dependencies() {
+    Get.put(HomeController());
+  }
+}
