--- lib/bindings/inductions_binding.dart
@@ -1,9 +1,0 @@
-import 'package:get/instance_manager.dart';
-import 'package:nittfest/controllers/inductions_controller.dart';
-
-class InductionsBindings extends Bindings {
-  @override
-  void dependencies() {
-    Get.put(InductionsController());
-  }
-}
