+++ lib/views/pages/home/home_page.dart
@@ -2,12 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:get/get_state_manager/src/simple/get_responsive.dart';
 import 'package:get/get_state_manager/src/simple/get_view.dart';
 import 'package:nittfest/constants/break_points.dart';
-import 'package:nittfest/controllers/inductions_controller.dart';
+import 'package:nittfest/controllers/home_controller.dart';
 import 'package:nittfest/views/pages/home/responsive/home_page_desktop.dart';
 import 'package:nittfest/views/pages/home/responsive/home_page_phone.dart';
 import 'package:nittfest/views/pages/home/widgets/social_icons.dart';
 
-class HomePage extends GetView<InductionsController> {
+class HomePage extends GetView<HomeController> {
   const HomePage({Key? key}) : super(key: key);
 
   @override
@@ -27,7 +27,7 @@ class HomePage extends GetView<InductionsController> {
   }
 }
 
-class HomeResponsive extends GetResponsiveView<InductionsController> {
+class HomeResponsive extends GetResponsiveView<HomeController> {
   HomeResponsive({Key? key})
       : super(key: key, settings: BreakPoints.getResponsiveScreenSettings());
