+++ lib/views/pages/home/widgets/baloon.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:get/get_state_manager/src/rx_flutter/rx_obx_widget.dart';
 import 'package:get/instance_manager.dart';
-import 'package:nittfest/controllers/inductions_controller.dart';
+import 'package:nittfest/controllers/home_controller.dart';
 
 class Baloon extends StatelessWidget {
   const Baloon({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    var controller = Get.find<InductionsController>();
+    var controller = Get.find<HomeController>();
     Size size = MediaQuery.of(context).size;
     return Obx(() => MouseRegion(
         onEnter: (event) => controller.isHovered.value = true,
