+++ lib/views/routes/navigation_pages.dart
@@ -3,7 +3,7 @@ import 'package:nittfest/bindings/forms_binding.dart';
 import 'package:nittfest/views/pages/forms/forms_page.dart';
 import 'package:nittfest/views/pages/home/home_page.dart';
 import 'package:nittfest/views/routes/navigation_routes.dart';
-import 'package:nittfest/bindings/inductions_binding.dart';
+import 'package:nittfest/bindings/home_binding.dart';
 import 'package:nittfest/bindings/splash_binding.dart';
 import 'package:nittfest/views/pages/splash/splash_page.dart';
 
@@ -17,7 +17,7 @@ class NavigationPages {
         GetPage(
           name: NavigationRoutes.inductionsHomeRoute,
           page: () => const HomePage(),
-          binding: InductionsBindings(),
+          binding: HomeBindings(),
         ),
         GetPage(
           name: NavigationRoutes.inductionsFormsRoute,
